commit fixed from branch n23
